Implementation Restart
I want to practice my more deliberate
software engineering practices. Testing, performance,
multicore, all that stuff. I'm hoping this skiplist
implementation will be a nice small test-bed for
that, so I've moved the old contents into
"old", and the new implementation is in the root directory.

This was just about 2 hours or so of sketching out the
basic machinery and fixing the obvious typos.