Benchmark 2 (insert elements in decreasing order)
I added a new benchmark that adds on a strictly-smaller
element each time to the list.

I would expect that this is essentially an append, and
so as a result our performance would be much better than
the (presumably tree-based) std::set. However, according
to my initial (very clumsy) measurements, that's not the
case. Interesting! More investigation is needed.