Experiment From Benchmark 2
Implements a new experiment (see verify_theoretical_properties.cpp)
confirming that adding a new smallest element still induces O(logn)
comparisons.

A moment's thought: "it's obvious". So I have begun playing around
with the idea of replacing, instead of a linear search, a binary
search at each node. My intuition is that there will be some size
cutoff when we should go back to linear search, but this will be
worth experimenting with.

Another observation: experimentally, the number of comparisons for
inserting a new smallest element tracks very well with 1*logn. But
the comparisons for inserting is 2*logn or 3*logn. Do I have an
off-by-one error? That would be great for improving performance.