Random Tests With Fixes
I created a test that randomly chooses heights and values
to insert into a skip-list. Motivated by memory leaks
and crashes when I tried to create a similar benchmark
for performance tests.

This flushed out 3 bugs: 1 in insertion (basically, we could
add the smallest element repeatedly), and 2 bugs in deletion.
The fundamental issue is a confusion about how when a predecessor
slice at index i is nullptr, that still means we need to check
if heads[i] has relevant data.

That may be a real implementation flaw, in as much as it has
confused me a few times. For now, I will continue to fill out
correctness testing.