Insertion and erasure cleanup
This contains a refactoring of insertion and removal.
An error_test was added to address a bug in erasure.
This also has a commented-out start of a binary search
implementation. It also contains a few small bugfixes
about iteration, returning references rather than copies,
etc.

So far, that binary search implementation increases
the number of comparisons, rather than decreases.
I only used it to find the first non-null element
to iterate over, so it's not too surprising.

Still, the code is slowly converging and cleaning up.