Added initial exhaustive test
I added an initial "exhaustive testing" test file.

There are two input sequences: the first is the sequence
of insertions/deletes to do. A positive number means
insert that number, and a negative number means remove
the abs of that number. If the largest absolute value allowed
is 4, then we have 8^N possible input sequences of length N.
(We exclude 0.)

There is a similar sequence for heights.

I would like for N = 7 and the maximum number to be 4, but that
takes a very long time to run. I reduced it to 5, and it takes about
10 seconds. Not ideal, but now I'm happy with that "exhaustive"
testing and am that much more confident in its correctness.

As always, this exhaustive testing stuff could also lead to
nice benchmarking opportunities.